Move capturer classes to 'capturer' package
